Refactor memory block to separate source
